pyproject.toml
[project]
name = "copick-tomotwin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "copick-utils",
    "copick>=0.8.1",
    "cryoet-data-portal>=4.0.0",
    "matplotlib>=3.10.1",
    "mrcfile>=1.5.4",
    "mysql-connector-python>=9.2.0",
    "pandas>=2.2.3",
    "tables>=3.10.2",
    "tomotwin-cryoet>=0.9.1.2",
    "torch==2.5.1",
    "pyarrow>=19.0.1",
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "pytest", # testing
    "mypy",  # linting
    "ruff",  # linting
]

[tool.uv.sources]
copick-utils = { git = "https://github.com/copick/copick-utils.git" }
